fix(#3376): extract shared path validator into app/utils/path_guard
